FEATURE: Automate workspace handling for reviews
The review workspaces for the users have been created automatically all the time. This change will bring some convenience. When we create a new backend user, we create a review workspace for the user on the fly.

When the user is a restricted editor, the user is not allowed to publish to live, but has no other workspace where publishing is possible. So we change the base workspace to the own review workspace. For non-restricted editors, we don’t switch the workspace.

If the user has never logged in before, we create a personal workspace with the review workspace as the base.

We also refacor the settings a bit and remove the workspace related functions from the NotifierService. So the NotifierService is just sending slack messages.